+++ packages/Webkul/Admin/src/Http/Controllers/Catalog/CategoryController.php
@@ -6,9 +6,9 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Event;
 use Webkul\Admin\Http\Controllers\Controller;
 use Webkul\Core\Repositories\ChannelRepository;
-use Webkul\Core\Http\Requests\MassUpdateRequest;
-use Webkul\Core\Http\Requests\MassDestroyRequest;
-use Webkul\Category\Http\Requests\CategoryRequest;
+use Webkul\Admin\Http\Requests\MassUpdateRequest;
+use Webkul\Admin\Http\Requests\MassDestroyRequest;
+use Webkul\Admin\Http\Requests\CategoryRequest;
 use Webkul\Admin\DataGrids\Catalog\CategoryDataGrid;
 use Webkul\Category\Repositories\CategoryRepository;
 use Webkul\Attribute\Repositories\AttributeRepository;
