+++ packages/Webkul/Admin/src/Http/Controllers/Catalog/ProductController.php
@@ -7,8 +7,8 @@ use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Arr;
 use Webkul\Admin\Http\Controllers\Controller;
-use Webkul\Product\Http\Requests\InventoryRequest;
-use Webkul\Product\Http\Requests\ProductForm;
+use Webkul\Admin\Http\Requests\InventoryRequest;
+use Webkul\Admin\Http\Requests\ProductForm;
 use Webkul\Category\Repositories\CategoryRepository;
 use Webkul\Attribute\Repositories\AttributeFamilyRepository;
 use Webkul\Inventory\Repositories\InventorySourceRepository;
@@ -19,8 +19,8 @@ use Webkul\Product\Repositories\ProductDownloadableSampleRepository;
 use Webkul\Product\Repositories\ProductInventoryRepository;
 use Webkul\Admin\Http\Resources\AttributeResource;
 use Webkul\Admin\DataGrids\Catalog\ProductDataGrid;
-use Webkul\Core\Http\Requests\MassUpdateRequest;
-use Webkul\Core\Http\Requests\MassDestroyRequest;
+use Webkul\Admin\Http\Requests\MassUpdateRequest;
+use Webkul\Admin\Http\Requests\MassDestroyRequest;
 use Webkul\Core\Rules\Slug;
 use Webkul\Product\Helpers\ProductType;
 use Webkul\Product\Facades\ProductImage;
