+++ packages/Webkul/Admin/src/Http/Controllers/Customers/CustomerController.php
@@ -9,8 +9,8 @@ use Webkul\Admin\Http\Controllers\Controller;
 use Webkul\Customer\Repositories\CustomerRepository;
 use Webkul\Customer\Repositories\CustomerGroupRepository;
 use Webkul\Admin\DataGrids\Customers\CustomerDataGrid;
-use Webkul\Core\Http\Requests\MassUpdateRequest;
-use Webkul\Core\Http\Requests\MassDestroyRequest;
+use Webkul\Admin\Http\Requests\MassUpdateRequest;
+use Webkul\Admin\Http\Requests\MassDestroyRequest;
 use Webkul\Customer\Repositories\CustomerNoteRepository;
 
 class CustomerController extends Controller
