+++ packages/Webkul/Admin/src/Http/Controllers/Marketing/Promotions/CartRuleCouponController.php
@@ -4,7 +4,7 @@ namespace Webkul\Admin\Http\Controllers\Marketing\Promotions;
 
 use Illuminate\Http\JsonResponse;
 use Webkul\Admin\Http\Controllers\Controller;
-use Webkul\Core\Http\Requests\MassDestroyRequest;
+use Webkul\Admin\Http\Requests\MassDestroyRequest;
 use Webkul\CartRule\Repositories\CartRuleCouponRepository;
 use Webkul\Admin\DataGrids\Marketing\Promotions\CartRuleCouponDataGrid;
