+++ packages/Webkul/Admin/src/Http/Controllers/Settings/UserController.php
@@ -10,7 +10,7 @@ use Illuminate\Support\Facades\Storage;
 use Webkul\Admin\Http\Controllers\Controller;
 use Webkul\User\Repositories\AdminRepository;
 use Webkul\User\Repositories\RoleRepository;
-use Webkul\User\Http\Requests\UserForm;
+use Webkul\Admin\Http\Requests\UserForm;
 use Webkul\Admin\DataGrids\Settings\UserDataGrid;
 
 class UserController extends Controller
