+++ packages/Webkul/Admin/src/Http/Requests/CatalogRuleRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Webkul\CatalogRule\Http\Requests;
+namespace Webkul\Admin\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
